Remove cycle between ReactFiberHooks and ReactInternalTypes (#20242)
